mm-rag-chat/requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=1.9.0
transformers>=4.11.3
gradio>=3.0.1

# Google Cloud
google-cloud-core>=2.3.2
google-cloud-logging>=3.5.0
google-cloud-storage>=2.7.0
google-cloud-aiplatform>=1.25.0
google-auth>=2.22.0

# Database
mysql-connector-python>=8.0.26
pinecone-client>=2.2.1

# Vector embeddings and image processing
sentence-transformers>=2.2.0
Pillow>=8.3.1

# Utilities
requests>=2.26.0
python-dotenv>=0.19.0

# Optional: for development
pytest>=6.2.5
flake8>=3.9.2
black>=21.7b0

# Additional potentially useful libraries
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
